feat: ingress rule restructuring completed
